Removing redundent files
